A class with public static methods
Transport class obtains number of routes from user for each item, then it obtains how many miles user like to travel. According to the miles, it suggest to user which type of vehicle user should be use for the journey. For example, if user like like travel only 5 miles then it suggest Bike.